text_analyzer/requirements.txt
Flask
nltk
spacy
textstat
pyspellchecker
matplotlib
rake-nltk
